requirements.txt
requests>=2.28.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0

# HTML/Playwright generation (default method)
playwright>=1.40.0

# Web UI
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
jinja2>=3.1.0

# Optional - for Gemini AI provider
# google-genai>=0.4.0
